requirements.txt
python-pptx>=0.6.23
mistune>=3.0.0
click>=8.1.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
